requirements.txt
attrs==20.3.0
importlib-metadata==3.4.0
iniconfig==1.1.1
numpy==1.20.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-json-logger==2.0.1
PyYAML==5.4.1
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.0
